Update README.md
I fixed a couple wording errors and added (REWORD) to a couple lines in the System Requirements section. Those lines didn't quite make sense. Wording changes could be made to make the Getting Started and Example section more friendly. I think it is to the point now, but that's not always a bad thing. Consider rewording if you have time, but I think it's shippable at this point.